refactor: Fix all the imports in moved tests
