+++ services/config-server/src/services/ValidatorService.test.ts
@@ -1,5 +1,5 @@
 import { ConfigFileType } from "@powerpi/common";
-import ValidatorService from "../../src/services/ValidatorService";
+import ValidatorService from "./ValidatorService";
 import { setupValidator, testInvalid } from "./tests/setupValidator";
 
 describe("ValidatorService", () => {
